+++ scripts/data/FilterMixin.mjs
@@ -2,9 +2,13 @@ import {BabonusTypes} from "../applications/dataModel.mjs";
 
 export function FilterMixin(Base) {
   return class BaseFilter extends Base {
+    // The name of this filter.
     static name = null;
+    // Whether this filter can be added more than once to a babonus.
     static repeatable = false;
+    // What handlebars template to use when rendering this filter in the builder.
     static template = null;
+    // Whether this filter has 'exclude' as an option in KeysDialog.
     static canExclude = false;
 
     /**
@@ -52,8 +56,25 @@ export function FilterMixin(Base) {
      * @param {Babonus} bonus     The babonus about to be saved.
      * @returns {boolean}         Whether to save the filter.
      */
-    static storage(bonus){
+    static storage(bonus) {
       return this.value(bonus).length > 0;
+    }
+
+    /**
+     * Return an array objects with 'value' and 'label', related to what this field should show.
+     * @returns {object[]}
+     */
+    static get choices() {
+      throw new Error("This must be subclassed!");
+    }
+
+    /**
+     * Helper method for creating an array of choices.
+     * @returns {string[]}      An array of string options.
+     */
+    static get stringOptions() {
+      if (this.canExclude) return this.choices.flatMap(({value}) => [value, `!${value}`]);
+      else return this.choices.map(({value}) => value);
     }
   }
 }
