+++ scripts/data/fields/HealthPercentagesField.mjs
@@ -1,4 +1,4 @@
-import {FilterMixin} from "../../FilterMixin.mjs";
+import {FilterMixin} from "../FilterMixin.mjs";
 
 export class HealthPercentagesField extends FilterMixin(foundry.data.fields.SchemaField) {
   static name = "healthPercentages";
